refactor: move PinnedTabsScaffold.kt to scaffold directory
